pyproject.toml
[project]
name = "gitgossip"
version = "0.1.0"
description = "AI that spills the tea on your commits — human-readable Git summaries."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Osman Goni Nahid" }
]

# Core runtime dependencies
dependencies = [
    "gitpython>=3.1.45",
    "openai>=2.3.0",
    "psutil>=7.1.0",
    "pydantic==2.8.0",
    "rich>=14.1.0",
    "ruamel-yaml>=0.18.15",
    "typer>=0.9.0",
]

# CLI entrypoint definition
[project.scripts]
gitgossip = "gitgossip.cli:app"

# Required by PyPI and uv to treat as installable package
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# uv-specific configuration
[tool.uv]
package = true
dev-dependencies = [
    "pytest",
    "black",
    "isort",
    "pylint>=3.3.9",
    "mypy>=1.18.2",
    "ruff>=0.14.0",
    "pre-commit>=4.3.0",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
extend-exclude = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"


